* commit 'c70f214d8f48d1c1cafd230d1ab18b1c31ab8cdd':
  Fix sign comparison warnings
  Fix out-of-order variable initialization
  Remove unused local variables
  Fix pessimistic move warning
  Fix sign compare warning in gtest
  Fix unused params on release build
  Fix different sign comparisons
  Fix r3d missing header warning
  Ignore Clion files
  Fix untriggered exception
  Fix missing return variable

+++ wonton/mesh/direct_product/test/test_direct_product_mesh_wrapper.cc
@@ -38,7 +38,7 @@ namespace direct_product_mesh_wrapper_test {
 
     // For each axis
     for (int d = 0; d < D; ++d) {
-      ASSERT_EQ(mesh_wrapper.axis_num_cells(d), axis_points[d].size()-1);
+      ASSERT_EQ(mesh_wrapper.axis_num_cells(d), (int) axis_points[d].size()-1);
       int n = 0;
       for (auto iter = mesh_wrapper.axis_point_begin(d);
            iter != mesh_wrapper.axis_point_end(d);
